feat: 에디터 컨텍스트 내에 SubmitHeader 추가 (#35)
